Fix multiple_fibers_cubes_partition

+++ core/src/time_stepping_scheme/time_stepping_scheme.cpp
@@ -122,7 +122,7 @@ double TimeSteppingScheme::endTime()
   return endTime_;
 }
 
-double TimeSteppingScheme::numberTimeSteps()
+int TimeSteppingScheme::numberTimeSteps()
 {
   return numberTimeSteps_;
 }
